some bugs are fixed

+++ src/main/java/com/tischenko/models/analyzers/poliz/Id.java
@@ -1,7 +1,7 @@
 package com.tischenko.models.analyzers.poliz;
 
 public class Id implements Valuable, ProgramPart {
- private double value = 0;
+ private int value = 0;
  private String name;
 
   public Id(String name) {
@@ -14,7 +14,7 @@ public class Id implements Valuable, ProgramPart {
   }
 
   public void setValue(double value) {
-    this.value = value;
+    this.value = (int)value;
   }
 
   public String getName() {
